third_party/blink/web_tests/external/wpt/workers/dedicated-worker-parse-error-failure-expected.txt

This is a testharness.js-based test.
FAIL Classic worker construction for script with syntax error should dispatch an event named error. assert_equals: expected function "function Event() { [native code] }" but got function "function ErrorEvent() { [native code] }"
FAIL Static import on classic worker should dispatch an event named error. assert_equals: expected function "function Event() { [native code] }" but got function "function ErrorEvent() { [native code] }"
Harness: the test ran to completion.
